View commits made between two commits

compare_commits() summarises the commits made between two commits. It returns a list of statistics along with a table of the commits with the properties as columns and a row for each commit. The base commit must be in the history of the head commit.

compare_commits(base, head, repo, ...)

Arguments

base
(string) Either a SHA, branch or tag used to identify the base commit.

head
(string) Either a SHA, branch or tag used to identify the head commit.

repo
(string) The repository specified in the format: owner/repo.

...
Parameters passed to gh_request().

Value

compare_commits() returns a list of comparison properties.

Comparison Properties:

status: The status of the comparison, can be "ahead" "behind" or "diverged".

ahead_by: The number of commits the head commit is ahead of the base commit.

behind_by: The number of commits the head commit is behind the base commit.

total_commits: The total number of commits between the head commit and the base commit.

html_url: The address of the commit comparison web page.

commits: A tibble summarising the commits between the head commit and the base commit.

Commit Properties:

sha: The commit SHA.

message: The commit message.

author_login: The author's account login.

author_name: The author's name.

author_email: The author's email address.

committer_login: The committer's account login.

committer_name: The committer's name.

committer_email: The committer's email address.

tree_sha: The SHA of the commit's file tree.

parents: The SHAs of the parent commits.

date: The time and date the commit was made.

html_url: The address of the commit's web page.

Details

For more details see the GitHub API documentation:

https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#compare-two-commits

Examples

if (FALSE) {

  # View the changes made between the current main branch and a release
  compare_commits("main", "0.8.7", "ChadGoymer/githapi")

}