Create a new files

This function creates a new text files in a GitHub repository. It adds a new commit for each file created with the specified commit message and on the specified branch. The committer and author can be set explicitly if necessary.

create_files(
  paths,
  contents,
  messages,
  branches,
  parents,
  committer,
  author,
  repo,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)

Arguments

paths
(character) The paths to the files in the repository.

contents
(character) The text to add to the files.

messages
(character) The commit messages to use. If a single message is provided it is used for all commits with the file added as a suffix.

branches
(character, optional) The branches to make the commits on. If a single branch is provided all commits are made on it. If not specified the default branch for the repository is used.

parents
(character, optional) SHAs, branches or tags of the commits to use as parents if creating a new branch. Not required if making a new commit on an existing branch. A character vector can be supplied corresponding to a different parent for each file, or a string can be provided if the same parent is used for all files.

committer
(list, optional) The name and email address of the committer. This needs to be specified as a named list, e.g. list(name = "Bob Smith", email = "bob.smith@acme.com"). If not specified then the authenticated user is used.

author
(list, optional) The name and email address of the author. This needs to be specified as a named list, e.g. list(name = "Bob Smith", email = "bob.smith@acme.com"). If not specified then the authenticated user is used.

repo
(string) The repository specified in the format: owner/repo.

token
(string, optional) The personal access token for GitHub authorisation. Default: value stored in the environment variable GITHUB_TOKEN (or GITHUB_PAT) or in the R option "github.token".

api
(string, optional) The URL of GitHub's API. Default: the value stored in the environment variable GITHUB_API or in the R option "github.api".

...
Parameters passed to gh_request().

Value

A tibble describing the files, with the following columns (see GitHub's documentation for more details):

name: The name of the file.

path: The path to the file.

sha: The SHA of the file.

size: The size of the file in bytes.

type: Either "file", "dir", "symlink" or "submodule".

url: The URL to the file contents.

html_url: The URL to view the file in GitHub.

git_url: The URL to the Git blob.

download_url: The URL to download the file.

commit_message: The commit message recorded in GitHub.

commit_sha: The SHA of the commit.

commit_url: The URL to the commit.

commit_author: The author of the commit.

commit_committer: The committer of the commit.

commit_tree_sha: The SHA for the git tree.

commit_tree_url: The URL for the git tree.

commit_parent_sha: The SHA of the parent commit.

commit_parent_url: The URL to the parent commit.

Details

https://developer.github.com/v3/repos/contents/#create-a-file