Create a milestone in a repository

This function creates a new milestone for the specified repository in GitHub. It can also be used to set a due date.

create_milestone(title, repo, description, due_on, state = "open", ...)

Arguments

title
(string) The title of the milestone.

repo
(string) The repository specified in the format: owner/repo.

description
(string, optional) A description of the milestone.

due_on
(string, optional) The milestone due date. This is in the format: YYYY-MM-DD.

state
(string, optional) The state of the milestone. Either "open" or "closed". Default: "open".

...
Parameters passed to gh_request().

Value

create_milestone() returns a list of the milestone's properties.

Milestone Properties:

number: The number assigned to the milestone within the repository.

title: The title of the milestone.

description: The description of the milestone.

state: The state of the milestone - either "open" or "closed".

open_issues: The number of open issues within the milestone.

closed_issues: The number of closed issues within the milestone.

html_url: The URL of the milestone's web page in GitHub.

creator: The creator's login.

created_at: When the milestone was created.

updated_at: When the milestone was last updated.

due_on: When the milestone is due.

closed_at: When the milestone was closed.

Details

For more details see the GitHub API documentation:

https://developer.github.com/v3/issues/milestones/#create-a-milestone

Examples

if (FALSE) {

  create_milestone(
    title       = "test milestone",
    repo        = "ChadGoymer/githapi",
    description = "This is a test milestone",
    due_on      = "2030-01-01 00:00:00")

}