Create a release in a repository

This function creates a new release in the specified repository in GitHub. It must be pointed at a commit by providing an existing tag or a Git reference, which can be either a SHA or branch. If the tag does not exist it is created pointing at the commit; if it does exist then the reference is ignored.

create_release(
  tag,
  repo,
  name,
  body,
  ref,
  draft = FALSE,
  prerelease = FALSE,
  ...
)

Arguments

tag
(string) The name of the tag.

repo
(string) The repository specified in the format: owner/repo.

name
(string, optional) The name of the release.

body
(string, optional) The description of the release.

ref
(string, optional) Either a SHA or branch used to identify the commit. If the tag exists this ref is ignored. If not supplied the repositories default branch is used.

draft
(boolean, optional) Whether the release is a draft. Default: FALSE.

prerelease
(boolean, optional) Whether the release is a pre-release. Default: FALSE.

...
Parameters passed to gh_request().

Value

create_release() returns a list of the release's properties.

Release Properties:

id: The id of the release.

tag: The tags associated with the release.

name: The name of the release.

body: The description of the release.

commit: The commit associated with the release.

draft: Whether the release is draft.

prerelease: Whether it is a pre-release.

author_login: The author's account login.

assets: The name of the assets associated with the release.

html_url: The address of the release's web page.

created_at: The time and date the release was created.

published_at: The time and date the release was published.

Details

For more details see the GitHub API documentation:

https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#create-a-release

Examples

if (FALSE) {

  # Create a release on the default branch
  create_release(
    tag  = "1.0.0",
    repo = "ChadGoymer/githapi",
    name = "Initial production release",
    body = "This is a release created by create_release()"
  )

  # Create a draft pre-release
  create_release(
    tag        = "1.0.9000",
    repo       = "ChadGoymer/githapi",
    name       = "Draft 1.1 release",
    body       = "This is a release created by create_release()",
    ref        = "dev-branch",
    draft      = TRUE,
    prerelease = TRUE
  )

}