Create a status for a commit

This function creates a new status for a commit in the specified repository in GitHub. The status can also have a description and a target URL, allowing a user to link to more information. The context allows you to differentiate the status from others. The status can then be updated, from pending to success say, by using the same context.

create_status(
  state,
  ref,
  repo,
  description,
  target_url,
  context = "default",
  ...
)

Arguments

state
(string) The state of the status. Can be one of "error", "failure", "pending", or "success".

ref
(string) Either a SHA, branch or tag used to identify the commit.

repo
(string) The repository specified in the format: owner/repo.

description
(string, optional) A short description of the status.

target_url
(string, optional) The target URL to associate with this status. This URL will be linked from the GitHub UI to allow users to easily see the source of the status.

context
(string, optional) A string label to differentiate this status from the status of other systems. Default: "default".

...
Parameters passed to gh_request().

Value

create_status() returns a list of the status's properties.

Status Properties:

id: The id of the release.

state: The state of the status.

description: The description of the status.

target_url: The URL linked to the status.

context: The context of the status.

creator: The creator of the status.

created_at: When it was created.

updated_at: When it was last updated.

Details

For more details see the GitHub API documentation:

https://docs.github.com/en/rest/reference/repos#create-a-commit-status

Examples

if (FALSE) {

  # Create a status on the main branch
  create_status(
    state       = "pending",
    ref         = "main",
    repo        = "ChadGoymer/test-githapi",
    description = "This is a pending status",
    target_url  = "https://www.goymer.me.uk/githapi",
    context     = "test/githapi")

  # Update the status on the main branch
  create_status(
    state       = "success",
    ref         = "main",
    repo        = "ChadGoymer/test-githapi",
    description = "This is a success status",
    target_url  = "https://www.goymer.me.uk/githapi",
    context     = "test/githapi")

}