Create tags at specified commits.

This function creates the specified tags at the specified commits within a repository in GitHub.

create_tags(repo, tags, shas, token = getOption("github.token"),
  api = getOption("github.api"), ...)

Arguments

repo
(string) The repository specified in the format: owner/repo.

tags
(character) The tag names.

shas
(character) The SHAs of the commits to tag.

token
(string, optional) The personal access token for GitHub authorisation. Default: value stored in the environment variable GITHUB_TOKEN (or GITHUB_PAT) or in the R option "github.token".

api
(string, optional) The URL of GitHub's API. Default: the value stored in the environment variable GITHUB_API or in the R option "github.api".

...
Parameters passed to gh_request().

Value

A tibble describing the tags, with the following columns (see GitHub's documentation for more details):

name: The name of the tag.

ref: The full git reference for the tag.

url: The URL to get the tag details from GitHub.

object_sha: The SHA of the object tagged.

object_type: The type of the object tagged.

object_url: The URL to get the object details from GitHub

Details

https://developer.github.com/v3/git/refs/#create-a-reference