Create a team in a GitHub organization

This function creates a new team in the specified organization in GitHub. It can also be used to specify the maintainers, the repositories to add the team to and a parent team.

create_team(
  name,
  org,
  description,
  maintainers,
  repo_names,
  privacy,
  parent_team,
  ...
)

Arguments

name
(string) The name of the team.

org
(string) The login of the organization.

description
(string, optional) The description of the team.

maintainers
(character, optional) The logins of organization members to add as maintainers of the team. If you do not specify any maintainers, then you will automatically become a team maintainer when you create a new team.

repo_names
(character, optional) The full name (e.g. "organization-name/repository-name") of repositories to add the team to.

privacy
(string, optional) The level of privacy this team should have. The options are: For a non-nested team:

"secret": only visible to organization owners and members of this team.

"closed": visible to all members of this organization.

Default: "secret".

For a parent or child team:

"closed": visible to all members of this organization.

Default for child team: "closed"

parent_team
(integer or string, optional) The ID or name of a team to set as the parent team.

...
Parameters passed to gh_request().

Value

create_team() returns a list of the team's properties.

Team Properties:

id: The ID of the team.

name: The name of the team.

organization: The organization the team is associated with.

slug: The team slug name.

description: The description of the team.

privacy: The privacy setting of the team - either "closed" or "secret".

permission: The default repository permissions of the team.

parent: The parent team.

members_count: The number of members.

repos_count: The number of repositories the team has access to.

created_at: When it was created.

updated_at: When it was last updated.

html_url: The URL of the team page in GitHub

Details

For more details see the GitHub API documentation:

https://developer.github.com/v3/teams/#create-team

Examples

if (FALSE) {

  # Create a team in an organization
  create_team("TestTeam", "HairyCoos")

  # Create a team and specify maintainers
  create_team("TestTeam2", "HairyCoos", maintainers = c("ChadGoymer", "chris-walker1"))

  # Create a team and set the privacy
  create_team("TestTeam3", "HairyCoos", privacy = "closed")

  # Create a team and specify a parent team
  create_team("TestTeam4", "HairyCoos", parent_team = "TestTeam3")

}