Get multiple pages from the GitHub API

This function is used when requesting a collection of entities. GitHub sets a maximum page size of 100, so if more are request multiple requests are made and the results are combined. Each page uses gh_request() to retrieve the contents.

gh_page(
  url,
  n_max = 1000,
  headers = NULL,
  accept = "application/vnd.github.v3+json",
  token = gh_token(),
  proxy = getOption("github.proxy"),
  ...
)

Arguments

url
(string) The address of the API endpoint.

n_max
(integer, optional) Maximum number to return. Default: 1000.

headers
(character, optional) Headers to add to the request. Default: NULL.

accept
(string, optional) The mime format to accept when making the call. Default: "application/vnd.github.v3+json".

token
(string or Token, optional) An authorisation token to include with the request. If NULL the OAuth process is triggered. Default: NULL.

proxy
(character, optional) The proxy server to use to connect to the github API. If NULL then no proxy is used. Can be set in the option github.proxy or the environment variable GITHUB_PROXY. Default: NULL.

...
Parameters passed to gh_request().

Value

A github list object consisting of the response, parsed into a list, with the attributes:

url: The URLs the request was sent to

request: The type of HTTP request made

status: The HTTP status code returned

header: The HTTP header returned

Examples

if (FALSE) {

  # First 20 users
  gh_page(
    url   = "https://api.github.com/users",
    n_max = 20
  )

  # First 150 users (two pages)
  gh_page(
    url   = "https://api.github.com/users",
    n_max = 150
  )

}