Get information about a user's, organization's or team's repositories.

https://developer.github.com/v3/repos/#list-user-repositories https://developer.github.com/v3/repos/#list-organization-repositories https://developer.github.com/v3/orgs/teams/#list-team-repos

gh_repositories(
  owner,
  team,
  type = NULL,
  sort = NULL,
  direction = NULL,
  n_max = 1000L,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)

Arguments

owner
(string) The user or organization owning the repository.

team
(string) The team owning the repository.

type
(string, optional) Can be one of all, owner, member. Default: owner.

sort
(string, optional) Can be one of created, updated, pushed, full_name. Default: full_name.

direction
(string, optional) Can be one of asc or desc. Default: when using full_name: asc, otherwise desc.

n_max
(integer, optional) Maximum number to return. Default: 1000.

token
(string, optional) The personal access token for GitHub authorisation. Default: value stored in the environment variable GITHUB_TOKEN or GITHUB_PAT.

api
(string, optional) The URL of GitHub's API. Default: the value stored in the environment variable GITHUB_API or https://api.github.com.

...
Parameters passed to gh_page().

Value

A tibble describing all the repositories a user or organization has (see GitHub's API documentation for more detail).

Details

Note: Must specify either owner or team, but not both.