User-friendly access to the GitHub API for R, consistent with the tidyverse

Provides a suite of functions which simplify working with GitHub's API. The functions have a consistent API and return either lists for single entities and tibbles for collections.

See also