Deprecated functions in package githapi.

The functions listed below are deprecated and will be removed in the near future. When possible, alternative functions with similar functionality are also mentioned.

Git Data Functions

gh_git_blob(): use view_blobs() instead.

gh_git_commit(): use view_commits() instead.

gh_git_reference(): use view_tags() or view_branches() instead.

gh_git_references(): use view_tags() or view_branches() instead.

is_tag(): use tags_exist() instead.

gh_git_tag(): is being removed

gh_git_tree(): replace with view_trees()

gh_save(): replace with download_files()

gh_source(): replace with source_files()

Repositories Functions

gh_tags(): replace with view_tags()

gh_branch(): replace with view_branches()

gh_branches(): replace with view_branches()

is_branch(): replace with branches_exist()

gh_commit(): replace with view_commits()

is_valid_sha(): replace with shas_exist()

gh_commit_sha(): replace with view_shas()

gh_commits(): replace with view_history()

gh_compare_commits(): replace with compare_commits()

gh_compare_files(): replace with compare_files()

gh_readme(): replace with read_files()

gh_contents(): replace with read_files()

gh_download(): replace with download_commit()

gh_release(): replace with view_releases()

gh_releases(): replace with view_releases()

gh_repository(): replace with view_repository()

gh_repositories(): replace with view_repositories()

is_repository(): is being removed

Projects Functions

gh_project(): replace with view_project()

gh_projects(): replace with view_projects()

gh_column(): replace with view_column()

gh_columns(): replace with view_columns()

gh_card(): replace with view_card()

gh_cards(): replace with view_cards()

Organizations, teams and users

gh_organization(): replace with view_organization()

gh_organizations(): replace with view_organizations()

gh_membership(): replace with view_membership()

gh_memberships(): replace with view_memberships()

gh_members(): replace with view_users()

gh_user(): replace with view_user()

gh_users(): replace with view_users()

gh_user_email(): is being removed. Use view_user()

gh_team(): replace with view_team()

gh_teams(): replace with view_teams()

is_member(): is being removed

is_manager(): is being removed

is_collaborator(): is being removed

gh_collaborators(): replace with view_collaborators()

gh_permissions(): replace with view_collaborator()