Update a card in a GitHub project

update_card() can be used to change a card's note in a project in GitHub or archive it. move_card() can be used to reorder the cards or move them to other columns.

update_card(card, note, archived, ...)

move_card(card, position, after, column, project, repo, user, org, ...)

Arguments

card
(integer) The ID of the card.

note
(string, optional) The new note for the card.

archived
(boolean, optional) Whether to archive the card.

...
Parameters passed to gh_request().

position
(string, optional) Either "top" or "bottom".

after
(integer, optional) An ID of another card to place this one after.

column
(integer or string, optional) Either the ID or name of the column.

project
(integer or string) Either the project number or name. Only required if moving the column.

repo
(string, optional) The repository specified in the format: owner/repo.

user
(string, optional) The login of the user.

org
(string, optional) The name of the organization.

Value

update_card() returns a list of the card properties.

Card Properties:

id: The ID of the card.

content_id: The ID of the issue or pull request.

note: The content of a note.

archived: Whether the card has been archived.

creator: The creator of the note.

created_at: When it was created.

updated_at: When it was last updated.

Details

You can update a card associated with either a repository, user or organization, by supplying them as an input, as long as you have appropriate permissions.

You can move a card by either specifying the position, either "top" or "bottom", by specifying another card to place it after, or by specifying a column to move it to.

For more details see the GitHub API documentation:

https://developer.github.com/v3/projects/cards/#update-a-project-card

https://developer.github.com/v3/projects/cards/#move-a-project-card

Examples

if (FALSE) {

  # Update a note in a card
  update_card(card = 123456, note = "This is an updated note")

  # Archive a card
  update_card(card = 654321, archived = TRUE)

  # Move a card to the top of a column
  move_card(card = 123456, position = "top")

  # Move a card after another card
  move_card(card  = 123456, after = 654321)

  # Move card to another column
  move_card(
    card     = 123456,
    position = "top",
    column   = "Test column 2",
    project  = "Test project",
    repo     = "ChadGoymer/githapi")

}