Update a comment in a repository

This function updates a comment for a gist, issue, pull request or commit. It can be used to replace the content of the comment. The repo argument is not required if a gist is specified.

update_comment(comment, body, gist, repo, type = "gist", ...)

Arguments

comment
(integer) The id of the comment.

body
(string) The content of the comment.

gist
(string, optional) The ID of the gist.

repo
(string) The repository specified in the format: owner/repo.

type
(string, optional) Whether the comment is for a "gist", "issue", "pull_request" or "commit". Default: "gist".

...
Parameters passed to gh_request().

Value

update_comment() returns a list of the comment properties.

Comment Properties:

id: The id of the comment.

body: The content of the comment.

commit: The commit associated with the comment (only for pull request or commit comments).

path: The path of the file to add the comment to (only for pull request or commit comments).

position: The line number in the file to attach the comment to (only for pull request or commit comments).

user: The comment author's account login.

html_url: The address of the comment's web page.

created_at: The time and date the comment was created.

published_at: The time and date the comment was published.

Details

For more details see the GitHub API documentation:

https://developer.github.com/v3/gists/comments/#edit-a-comment

https://developer.github.com/v3/issues/comments/#edit-a-comment

https://developer.github.com/v3/pulls/comments/#edit-a-comment

https://developer.github.com/v3/repos/comments/#update-a-commit-comment

Examples

if (FALSE) {

  # Update a gist comment
  update_comment(
    comment = 3281939,
    body    = "This comment has been updated by update_comment()",
    gist    = "8e5be270de9a88168372293a813543f9")

  # Update an issue comment
  update_comment(
    comment = 622980929,
    body    = "This comment has been updated by update_comment()",
    type    = "issue",
    repo    = "ChadGoymer/githapi")

  # Update an pull request comment
  update_comment(
    comment = 418979473,
    body    = "This comment has been updated by update_comment()",
    type    = "pull_request",
    repo    = "ChadGoymer/githapi")

  # Update a commit comment
  update_comment(
    comment = 38899533,
    body    = "This comment has been updated by update_comment()",
    type    = "commit",
    repo    = "ChadGoymer/githapi")

}