Update an issue in a repository

This function updates an issue for the specified repository in GitHub. It can be used to change the title, body, or assignees, it can also be used to replace labels and milestones, or to close the issue.

update_issue(
  issue,
  repo,
  title,
  body,
  assignees,
  labels,
  milestone,
  state,
  ...
)

Arguments

issue
(string or character) The number or title of the issue.

repo
(string) The repository specified in the format: owner/repo.

title
(string, optional) The title of the issue.

body
(string, optional) The contents of the issue.

assignees
(character, optional) Logins for Users to assign to this issue. NOTE: Only users with push access can set assignees for new issues. Setting to NULL clears all assignees.

labels
(character, optional) Labels to associate with this issue. NOTE: Only users with push access can set labels for new issues. Setting to NULL clears all labels.

milestone
(character or integer, optional) The title or number of the milestone to associate this issue with. NOTE: Only users with push access can set the milestone for new issues. Setting to NULL clears the current milestone.

state
(string, optional) The state of the issue. Either "open" or "closed".

...
Parameters passed to gh_request().

Value

update_issue() returns a list of the issue's properties.

Issue Properties:

number: The number assigned to the issue.

title: The title of the issue.

body: The body contents of the issue.

assignees: The users assigned to the issue.

labels: The labels attached to the issue.

milestone: The milestone assigned to the issue.

state: The state of the issue - either "open" or "closed".

repository: The repository the issue is in.

pull_request: Whether the issue is a pull request.

html_url: The URL of the issue's web page in GitHub.

creator: The creator's login.

created_at: When the issue was created.

updated_at: When the issue was last updated.

closed_at: When the issue was closed.

Details

For more details see the GitHub API documentation:

https://developer.github.com/v3/issues/#edit-an-issue

Examples

if (FALSE) {

  # Update the properties of a issue
  update_issue(
    issue   = "test issue",
    repo        = "ChadGoymer/githapi",
    title       = "updated test issue",
    description = "This is an updated test issue",
    due_on      = "2020-12-01")

  # Close a issue
  update_issue(
    issue = "updated test issue",
    repo      = "ChadGoymer/githapi",
    state     = "closed")

}