Update membership of an organization or team

This function can be used to invite a user into an organization or team, or update their role within the organization or team.

update_membership(user, org, team, role, ...)

Arguments

user
(string) The login of the user.

org
(string) The login of the organization.

team
(integer or string, optional) The ID or name of the team.

role
(string, optional) The role to give the user. For an organization this is either "member" or "admin", for a team it is either "member" or "maintainer".

...
Parameters passed to gh_request().

Value

update_memberships() returns a list of membership properties

Membership Properties:

user: The user login.

organization: The organization login.

team: The team name (team membership only).

role: The role of the user - for organizations it is either "member" or "admin", for teams it is either "member" or "maintainer".

state: The state of the membership - either "active" or "pending".

Details

Note: you can only invite or update a user if the authenticate user is an organization "owner" or a team "maintainer".

For more details see the GitHub API documentation:

https://developer.github.com/v3/orgs/members/#add-or-update-organization-membership

https://developer.github.com/v3/teams/members/#add-or-update-team-membership

Examples

if (FALSE) {

  # Invite a new user into an organization
  update_membership("ChadGoymer2", "HairyCoos")

  # Update a user's role in an organization
  update_membership("ChadGoymer2", "HairyCoos", role = "admin")

  # Invite a new user into a team
  update_membership("ChadGoymer2", "HairyCoos", "HeadCoos")

  # Update a user's role in a team
  update_membership("ChadGoymer2", "HairyCoos", "HeadCoos", role = "maintainer")

}