Update a milestone in a repository

This function updates a milestone for the specified repository in GitHub. It can be used to change title, description or due date, but can also be used to close the milestone.

update_milestone(milestone, repo, title, description, due_on, state, ...)

Arguments

milestone
(string or character) The number or title of the milestone.

repo
(string) The repository specified in the format: owner/repo.

title
(string, optional) The title of the milestone.

description
(string, optional) A description of the milestone.

due_on
(string, optional) The milestone due date. This is in the format: YYYY-MM-DD HH:MM:SS.

state
(string, optional) The state of the milestone. Either "open" or "closed".

...
Parameters passed to gh_request().

Value

update_milestone() returns a list of the milestone properties.

Milestone Properties:

number: The number assigned to the milestone within the repository.

title: The title of the milestone.

description: The description of the milestone.

state: The state of the milestone - either "open" or "closed".

open_issues: The number of open issues within the milestone.

closed_issues: The number of closed issues within the milestone.

html_url: The URL of the milestone's web page in GitHub.

creator: The creator's login.

created_at: When the milestone was created.

updated_at: When the milestone was last updated.

due_on: When the milestone is due.

closed_at: When the milestone was closed.

Details

For more details see the GitHub API documentation:

https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#update-a-milestone

Examples

if (FALSE) {

  # Update the properties of a milestone
  update_milestone(
    milestone   = "test milestone",
    repo        = "ChadGoymer/githapi",
    title       = "updated test milestone",
    description = "This is an updated test milestone",
    due_on      = "2020-12-01"
  )

  # Close a milestone
  update_milestone(
    milestone = "updated test milestone",
    repo      = "ChadGoymer/githapi",
    state     = "closed"
  )

}