Update an organization's properties in GitHub

This function updates an organization's properties in GitHub. You can only update an organization if you have the appropriate permissions.

update_organization(
  org,
  name,
  description,
  email,
  location,
  company,
  billing_email,
  has_organization_projects,
  has_repository_projects,
  default_repository_permission,
  members_can_create_repositories,
  ...
)

Arguments

org
(string) The login of the organization.

name
(string, optional) The shorthand name of the company.

description
(string, optional) The description of the company.

email
(string, optional) The publicly visible email address.

location
(string, optional) The location.

company
(string, optional) The company name.

billing_email
(string, optional) Billing email address. This address is not publicized.

has_organization_projects
(boolean, optional) Toggles whether an organization can use organization projects.

has_repository_projects
(boolean, optional) Toggles whether repositories that belong to the organization can use repository projects.

default_repository_permission
(string, optional) Default permission level members have for organization repositories:

"read": can pull, but not push to or administer this repository.

"write": can pull and push, but not administer this repository.

"admin": can pull, push, and administer this repository.

"none": no permissions granted by default.

Default: "read".

members_can_create_repositories
(boolean, optional) Toggles the ability of non-admin organization members to create repositories. Can be one of:

TRUE: all organization members can create repositories.

FALSE: only organization owners can create repositories.

Default: TRUE. Note: A parameter can override this parameter. See members_allowed_repository_creation_type for details.

...
Parameters passed to gh_request().

Value

update_organization() returns a list of the organization's properties.

Organization Properties:

id: The ID of the organization.

login: The login name of the organization.

name: The name of the organization.

description: The description of the organization.

company: The name of the associated company.

blog: The address for a blog.

location: The geographical location.

email: The email address for the organization.

is_verified: Whether the organization has been verified.

has_organization_projects: Whether the organization can have projects.

has_repository_projects: Whether the organization's repositories can have projects.

public_repos: The number of public repositories.

public_gists: The number of public gists.

html_url: The address for the organization's GitHub web page.

created_at: When the organization was created.

total_private_repos: The number of private repositories.

owned_private_repos: The number of owned private repositories.

private_gists: The number of private gists.

disk_usage: The total disk usage for the organization.

collaborators: The number of collaborators.

billing_email: The email address for billing.

plan_name: The name of the GitHub plan.

plan_space: The total space allocated for the plan.

plan_private_repos: The number of private repositories for the plan.

default_repository_settings: The default access for new repositories.

two_factor_requirement_enabled: Whether members require two-factor authentication.

members_can_create_repositories: Whether members can create repositories.

Details

For more details see the GitHub API documentation:

https://developer.github.com/v3/orgs/#edit-an-organization

Examples

if (FALSE) {

  # Update some of your organization properties
  update_organization(
    org                             = "HairyCoos",
    description                     = "We are the Hairy Coos!",
    location                        = "The Highlands",
    default_repository_permission   = "write",
    members_can_create_repositories = FALSE)

}