Update a GitHub project

This function updates a project in GitHub. It can be used to change the name and body, but can also be used to close the project, change permissions or add a team.

update_project(
  project,
  name,
  body,
  state,
  permission,
  private,
  repo,
  user,
  team,
  org,
  ...
)

Arguments

project
(integer or string) Either the project number or name.

name
(string, optional) The new name for the project.

body
(string, optional) The new description of the project.

state
(string, optional) The new state of the project, either "open" or "closed".

permission
(string, optional) The new team or organisation permissions for the project, either "read", "write", "admin" or "none". Note: applies to team and organization projects only.

private
(boolean, optional) Whether the project should be private. Note: applies to team and organization projects only.

repo
(string, optional) The repository specified in the format: owner/repo.

user
(string, optional) The login of the user.

team
(string or integer, optional) The team ID or name.

org
(string, optional) The name of the organization.

...
Parameters passed to gh_request().

Value

update_project() returns a list of the project properties.

Project Properties:

id: The ID of the project.

number: The number of the project for the repository, user or organization.

name: The name given to the project.

body: The description given to the project.

state: Whether the project is "open" or "closed".

private: Whether the project is private (organization and team projects only).

org_permissions: The default permission for organization members (organization and team projects only).

team_permissions: The default permission for team members (team projects only).

creator: The user who created the project.

created_at: When it was created.

updated_at: When it was last updated.

html_url: The URL to view the project.

Details

You can update a project associated with either a repository, user, team or organization, by supplying them as an input, as long as you have appropriate permissions. Supplying a team that does not already have access to the project adds them. If they have access, then the team's permissions can be changed with the permission argument.

For more details see the GitHub API documentation:

https://developer.github.com/v3/projects/#update-a-project

https://developer.github.com/v3/teams/#add-or-update-team-project

Examples

if (FALSE) {
  # Update the name of a project for a repository
  update_project(
    project = "Repo project",
    name    = "Updated repo project",
    body    = "This is an updated repository's project",
    repo    = "ChadGoymer/test-githapi")

  # Update the state of a project for a user
  update_project(
    name  = "User project",
    state = "closed",
    user  = "ChadGoymer")

  # Update the permissions of a project for an organization
  update_project(
    name       = "Org project",
    permission = "read",
    private    = TRUE,
    org        = "HairyCoos")

  # Add a team to the project
  update_project(
    project = "Org project",
    team    = "HeadCoos",
    org     = "HairyCoos")

  # Update the team's permissions on the project
  update_project(
    project    = "Org project",
    permission = "write",
    team       = "HeadCoos",
    org        = "HairyCoos")
}