Update a release in a repository

This function updates a release in the specified repository. It can be used to update properties, such as name or whether it is draft, or it can be used to point the release at a new commit by providing a Git reference, which can be either a SHA or branch.

update_release(
  release,
  repo,
  tag,
  name,
  body,
  ref = "master",
  draft = FALSE,
  prerelease = FALSE,
  ...
)

Arguments

release
(string) The id or current tag of the release.

repo
(string) The repository specified in the format: owner/repo.

tag
(string, optional) The name of the new tag.

name
(string, optional) The name of the release.

body
(string, optional) The description of the release.

ref
(string, optional) Either a SHA or branch used to identify the commit. Ignored if the tag exists this ref is ignored. Default: "master".

draft
(boolean, optional) Whether the release is a draft. Default: FALSE.

prerelease
(boolean, optional) Whether the release is a pre-release. Default: FALSE.

...
Parameters passed to gh_request().

Value

update_release() returns a list of the release properties.

Release Properties:

id: The id of the release.

tag: The tags associated with the release.

name: The name of the release.

body: The description of the release.

commit: The commit associated with the release.

draft: Whether the release is draft.

prerelease: Whether it is a pre-release.

author_login: The author's account login.

assets: The name of the assets associated with the release.

html_url: The address of the release's web page.

created_at: The time and date the release was created.

published_at: The time and date the release was published.

Details

For more details see the GitHub API documentation:

https://developer.github.com/v3/repos/releases/#edit-a-release

Examples

if (FALSE) {

  update_release(
    release = "1.0.0",
    repo    = "ChadGoymer/githapi",
    tag     = "1.0.1",
    name    = "Updated production release",
    body    = "This release has been updated by update_release()")

  update_release(
    release    = 1234567,
    tag        = "1.0.1",
    repo       = "ChadGoymer/githapi",
    name       = "Promoted draft release",
    body       = "This release has been updated by update_release()",
    draft      = FALSE,
    prerelease = FALSE)

}