Update releases to new commits.

This function updates an existing releases. It can update the names or bodies (description) of the releases. You can also change whether the releases are draft and whether they are pre-releases.

update_releases(
  tags,
  names,
  bodies,
  repo,
  draft = FALSE,
  prerelease = FALSE,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)

Arguments

tags
(character) The tag names.

names
(character, optional) The names of the releases. If missing or NA then the name is not changed. Default: NA.

bodies
(character, optional) The bodies for the releases. If missing or NA then the body is not changed. Default: NA.

repo
(string) The repository specified in the format: owner/repo.

draft
(logical, optional) Whether the releases are draft. Default: FALSE.

prerelease
(logical, optional) Whether the releases are a prerelease. Default: FALSE.

token
(string, optional) The personal access token for GitHub authorisation. Default: value stored in the environment variable GITHUB_TOKEN (or GITHUB_PAT) or in the R option "github.token".

api
(string, optional) The URL of GitHub's API. Default: the value stored in the environment variable GITHUB_API or in the R option "github.api".

...
Parameters passed to gh_request().

Value

A tibble describing the releases, with the following columns (see GitHub's documentation for more details):

id: The unique ID assigned by GitHub.

tag_name: The name of the tag associated with the release.

name: The name, or title, of the release.

body: The body, or description, of the release.

author_login: The author of the release.

draft: Whether the release is a draft.

prerelease: Whether the release is a pre-release.

target_commitish: The commit or branch the release was applied to.

created_at: The date and time of creation.

published_at: The date and time of publication.

assets: The names of any assets associated with the release.

zipball_url: The URL to download a zip file of the release commit.

url: The URL to view the release details on GitHub.

Details

https://developer.github.com/v3/repos/releases/#edit-a-release