Update a tag in a repository

This function updates a tag in the specified repository to point at a new commit. It must be pointed at a commit by providing a Git reference, which can be either a SHA, branch or tag. For a branch, the head commit is used.

update_tag(tag, ref, repo, force = FALSE, ...)

Arguments

tag
(string) The name of the tag.

ref
(string) Either a SHA, branch or tag used to identify the new commit.

repo
(string) The repository specified in the format: owner/repo.

force
(boolean, optional) Whether to force the update if it is not a simple fast-forward. Default: FALSE.

...
Parameters passed to gh_request().

Value

update_tag() returns a list of the tag properties.

Tag Properties:

name: The name of the tag.

ref: The full Git reference of the tag.

sha: The commit SHA the tag is pointing at.

Details

For more details see the GitHub API documentation:

https://docs.github.com/en/free-pro-team@latest/rest/reference/git#update-a-reference

Examples

if (FALSE) {

  update_tag(
    tag  = "new-tag",
    ref  = "6b7b5a090d47fd3ef495620513a3f80da2487b1d",
    repo = "ChadGoymer/githapi"
  )

}