Update a team's properties in GitHub

This function updates a team's properties in GitHub. You can only update a team if you have the appropriate permissions.

update_team(team, name, description, privacy, parent_team, org, ...)

Arguments

team
(integer or string) The Id or current name of the team.

name
(string, optional) The new name of the team.

description
(string, optional) The description of the team.

privacy
(string, optional) The level of privacy this team should have. Editing teams without specifying this parameter leaves privacy intact. The options are:

For a non-nested team:

"secret": only visible to organization owners and members of this team.

"closed": visible to all members of this organization.

For a parent or child team:

"closed": visible to all members of this organization.

parent_team
(integer or string, optional) The ID or name of a team to set as the parent team.

org
(string, optional) The organization the team is associated with. Not required if your are specifying team and/or parent_team set by ID.

...
Parameters passed to gh_request().

Value

update_team() returns a list of the team's properties.

Team Properties:

id: The ID of the team.

name: The name of the team.

organization: The organization the team is associated with.

slug: The team slug name.

description: The description of the team.

privacy: The privacy setting of the team - either "closed" or "secret".

permission: The default repository permissions of the team.

parent: The parent team.

members_count: The number of members.

repos_count: The number of repositories the team has access to.

created_at: When it was created.

updated_at: When it was last updated.

html_url: The URL of the team page in GitHub

Details

For more details see the GitHub API documentation:

https://developer.github.com/v3/teams/#edit-team

Examples

if (FALSE) {

  # Update a team
  update_team(
    team        = "TestTeam",
    description = "This is a test team",
    privacy     = "closed",
    parent_team = "TestTeam3",
    org         = "HairyCoos")

}