Update a team's permissions on a repository

update_team_repository() allows you to add or update a team's permission on a repository. remove_team_repository() removes the team's access to a repository.

update_team_repository(repo, team, org, permission = "pull", ...)

remove_team_repository(repo, team, org, ...)

Arguments

repo
(string) The repository specified in the format: owner/repo.

team
(string) The team name.

org
(string) The name of the organization.

permission
(string, optional) The permission to set for the team. Either: "pull", "push", "admin", "maintain" or "triage". Default: "pull".

...
Parameters passed to gh_request().

Value

update_team_repository() and remove_team_repository() returns a TRUE if successful.

Details

The team's permission can be set to:

"pull": Team members can pull from this repository.

"push": Team members can pull from and push to this repository.

"admin": Team members can pull from, push to and administer this repository.

"maintain": Team members can manage the repository without access to sensitive or destructive actions. Recommended for project managers. Only applies to repositories owned by organizations.

"triage": Team members can proactively manage issues and pull requests without write access. Recommended for contributors who triage a repository. Only applies to repositories owned by organizations.

For more details see the GitHub API documentation:

https://developer.github.com/v3/teams/#add-or-update-team-repository

https://developer.github.com/v3/teams/#remove-team-repository

Examples

if (FALSE) {

  # Add read access for the specified team
  update_team_repository(
    repo = "HairyCoos/test-repository",
    team = "test-team",
    org  = "HairyCoos")

  # Update team's permission to "maintain"
  update_team_repository(
    repo       = "HairyCoos/test-repository",
    team       = "test-team",
    org        = "HairyCoos",
    permission = "maintain")

  # Remove team's access
  remove_team_repository(
    repo = "HairyCoos/test-repository",
    team = "test-team",
    org  = "HairyCoos")

}