View branches within a repository

view_branches() summarises branches in a table with the properties as columns and a row for each branch in the repository. view_branch() returns a list of all properties for a single branch.

view_branches(repo, n_max = 1000, ...)

view_branch(branch, repo, ...)

Arguments

repo
(string) The repository specified in the format: owner/repo.

n_max
(integer, optional) Maximum number to return. Default: 1000.

...
Parameters passed to gh_page() or gh_request().

branch
(string) The name of the branch.

Value

view_branches() returns a tibble of branch properties. view_branch() returns a list of properties for a single branch.

Branch Properties:

name: The name of the branch.

ref: The full Git reference of the branch.

sha: The commit SHA the branch is pointing at.

Details

For more details see the GitHub API documentation:

https://developer.github.com/v3/git/refs/#list-matching-references

https://developer.github.com/v3/git/refs/#get-a-single-reference

Examples

if (FALSE) {

  # View all branches in a repository
  view_branches("ChadGoymer/githapi")

  # View a single branch
  view_label("new-branch", "ChadGoymer/githapi")

}