View cards within a GitHub project

view_cards() summarises cards in a table with the properties as columns and a row for each card in a column of a project. view_card() returns a list of all properties for a single card.

view_cards(column, project, repo, user, org, n_max = 1000, ...)

view_card(card, ...)

Arguments

column
(integer or string) Either the ID or name of the column.

project
(integer or string) Either the project number or name.

repo
(string, optional) The repository specified in the format: owner/repo.

user
(string, optional) The login of the user.

org
(string, optional) The name of the organization.

n_max
(integer, optional) Maximum number to return. Default: 1000.

...
Parameters passed to gh_page() or gh_request().

card
(integer) The ID of the card.

Value

view_cards() returns a tibble of card properties. view_card() returns a list of properties for a single card.

Card Properties:

id: The ID of the card.

content_id: The ID of the issue or pull request.

note: The content of a note.

archived: Whether the card has been archived.

creator: The creator of the note.

created_at: When it was created.

updated_at: When it was last updated.

Details

You can summarise all the cards of a project associated with either a repository, user or organization, by supplying them as an input.

For more details see the GitHub API documentation:

https://docs.github.com/en/free-pro-team@latest/rest/reference/projects#list-project-cards

https://docs.github.com/en/free-pro-team@latest/rest/reference/projects#get-a-project-card

Examples

if (FALSE) {

  # View cards in a repository project
  cards <- view_cards(
    column  = "Test cards",
    project = "Test cards",
    repo    = "ChadGoymer/githapi"
  )

  # View cards in a user's project
  cards <- view_cards(
    column  = "Test cards",
    project = "Test cards",
    user    = "ChadGoymer"
  )

  # View cards in an organization's project
  cards <- view_cards(
    column  = "Test cards",
    project = "Test cards",
    org     = "HairyCoos"
  )

  # View a single card
  view_card(card = 123456)

}