View comments in GitHub

view_comments() summarises comments in a table with the properties as columns and a row for each comment registered for the gist, issue, pull request or commit. view_comment() returns a list of all properties for a single comment. browse_comment() opens the web page for the comment in the default browser.

view_comments(gist, issue, pull_request, commit, repo, n_max = 1000, ...)

view_comment(comment, gist, repo, type = "gist", ...)

browse_comment(comment, repo, type, ...)

Arguments

gist
(string, optional) The ID of the gist.

issue
(integer, optional) The issue number.

pull_request
(integer, optional) The pull request number.

commit
(string, optional) Either a SHA or branch used to identify the commit.

repo
(string) The repository specified in the format: owner/repo.

n_max
(integer, optional) Maximum number to return. Default: 1000.

...
Parameters passed to gh_page() or gh_request().

comment
(integer) The id of the comment.

type
(string, optional) Whether the comment is for a "gist", "issue", "pull_request" or "commit". Default: "gist".

Value

view_comments() returns a tibble of comment properties. view_comment() returns a list of properties for a single comment. browse_comment() opens the default browser on the comment page and returns the URL.

Comment Properties:

id: The id of the comment.

body: The content of the comment.

commit: The commit associated with the comment (only for pull request or commit comments).

path: The path of the file to add the comment to (only for pull request or commit comments).

position: The line number in the file to attach the comment to (only for pull request or commit comments).

user: The comment author's account login.

html_url: The address of the comment's web page.

created_at: The time and date the comment was created.

published_at: The time and date the comment was published.

Details

For more details see the GitHub API documentation:

https://docs.github.com/en/free-pro-team@latest/rest/reference/gists#list-gist-comments

https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#list-issue-comments

https://docs.github.com/en/free-pro-team@latest/rest/reference/pulls#list-review-comments-on-a-pull-request

https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#list-commit-comments

https://docs.github.com/en/free-pro-team@latest/rest/reference/gists#get-a-gist-comment

https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#get-an-issue-comment

https://docs.github.com/en/free-pro-team@latest/rest/reference/pulls#get-a-review-comment-for-a-pull-request

https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#get-a-commit-comment

Examples

if (FALSE) {

  # View comments on a gist
  view_comments(gist = "8e5be270de9a88168372293a813543f9")

  # View comments on an issue
  view_comments(
    issue = "test issue",
    repo  = "ChadGoymer/githapi"
  )

  # View comments on a pull request
  view_comments(
    pull_request = "test pull request",
    repo         = "ChadGoymer/githapi"
  )

  # View comments on a commit
  view_comments(
    commit = "main",
    repo   = "ChadGoymer/githapi"
  )

  # View a single gist comment
  view_comment(
    comment = 622980929,
    gist    = "8e5be270de9a88168372293a813543f9"
  )

  # View a single issue comment
  view_comment(
    comment = 622980929,
    type    = "issue",
    repo    = "ChadGoymer/githapi"
  )

  # View a single pull request comment
  view_comment(
    comment = 418979473,
    type    = "pull_request",
    repo    = "ChadGoymer/githapi"
  )

  # View a single commit comment
  view_comment(
    comment = 38899533,
    type    = "commit",
    repo    = "ChadGoymer/githapi"
  )

  # Browse an issue comment
  browse_comment(
    comment = 622980929,
    type    = "issue",
    repo    = "ChadGoymer/githapi"
  )

  # Browse a pull request comment
  browse_comment(
    comment = 418979473,
    type    = "pull_request",
    repo    = "ChadGoymer/githapi"
  )

  # Browse a commit comment
  browse_comment(
    comment = 38899533,
    type    = "commit",
    repo    = "ChadGoymer/githapi"
  )

}