View gists in GitHub

view_gists() summarises a user's gists in a table with the properties as columns and a row for each gist. view_gist() returns a list of all properties for a single gist. browse_gist() opens the web page for the gist in the default browser.

view_gists(user, since, n_max = 1000, ...)

view_gist(gist, ...)

browse_gist(gist, ...)

Arguments

user
(string, optional) The login of the user. If not specified the authenticated user is used.

since
(string, optional) A date & time to filter by. Must be in the format: YYYY-MM-DD HH:MM:SS.

n_max
(integer, optional) Maximum number to return. Default: 1000.

...
Parameters passed to gh_page() or gh_request().

gist
(string) The id of the gist.

Value

view_gists() returns a tibble of gist properties. view_gist() returns a list of properties for a single gist. browse_gist opens the default browser on the gist page and returns the URL.

Gist Properties:

id: The id for the gist

description: The description of the gist

files: For view_gists() just the filenames are returned, for view_gist() a tibble of file properties is returned:

filename: The name of the file

type: The type of file

content: The file content

size: The size of the file in bytes

truncated: Whether the file content has been truncated

owner: The login if the gist's owner

public: Whether the gist is public

html_url: The address of the gist's web page.

created_at: The time and date the gist was created.

updated_at: The time and date the gist was last updated.

Details

When viewing gists, if a user is not specified the gists for the authenticated user are returned.

For more details see the GitHub API documentation:

https://developer.github.com/v3/gists/#list-gists-for-a-user

https://developer.github.com/v3/gists/#list-gists-for-the-authenticated-user

https://developer.github.com/v3/gists/#get-a-gist

Examples

if (FALSE) {

  # View the authenticated user's gists
  view_gists()

  # View a specific user's gists
  view_gists("ChadGoymer")

  # View a specific gist
  view_gist("806dca6b09a39e7b6326a0c8137583e6")

  # Browse a gist
  browse_gist("806dca6b09a39e7b6326a0c8137583e6")

}