View issues within a repository or organization

view_issues() summarises issues in a table with the properties as columns and a row for each issue in the repository or organization. view_issue() returns a list of all properties for a single issue. browse_issue() opens the web page for the issue in the default browser.

view_issues(
  repo,
  org,
  labels,
  milestone,
  since,
  state = "open",
  sort = "created",
  direction = "desc",
  n_max = 1000,
  ...
)

view_issue(issue, repo, ...)

browse_issue(issue, repo, ...)

Arguments

repo
(string, optional) The repository specified in the format: owner/repo.

org
(string, optional) The name of the organization.

labels
(character, optional) Label names to filter by.

milestone
(string or integer) Milestone number or title to filter by.

since
(string, optional) A date & time to filter by. Must be in the format: YYYY-MM-DD HH:MM:SS.

state
(string, optional) The state of the issues to return. Can be either "open", "closed", or "all". Default: "open".

sort
(string, optional) The property to order the returned issues by. Can be either "created", "updated", or "comments". Default: "created".

direction
(string, optional) The direction of the sort. Can be either "asc" or "desc". Default: "desc".

n_max
(integer, optional) Maximum number to return. Default: 1000.

...
Parameters passed to gh_page() or gh_request().

issue
(string or character) The number or title of the issue.

Value

view_issues() returns a tibble of issue properties. view_issue() returns a list of properties for a single issue. browse_issue() opens the default browser on the issue's page and returns the URL.

Issue Properties:

number: The number assigned to the issue.

title: The title of the issue.

body: The body contents of the issue.

assignees: The users assigned to the issue.

labels: The labels attached to the issue.

milestone: The milestone assigned to the issue.

state: The state of the issue - either "open" or "closed".

repository: The repository the issue is in.

pull_request: Whether the issue is a pull request.

html_url: The URL of the issue's web page in GitHub.

creator: The creator's login.

created_at: When the issue was created.

updated_at: When the issue was last updated.

closed_at: When the issue was closed.

Details

You can summarise all the issues in a repository or organization by specifying the arguments. If neither are specified then all the issues assigned to the authenticated user are returned. You can filter the issues based on the labels, milestone, whether they have been updated since a specified date or whether they are "open" or "closed". Finally, the order the results are returned can be controlled with sort and direction.

For more details see the GitHub API documentation:

https://developer.github.com/v3/issues/#list-issues

https://developer.github.com/v3/issues/#list-issues-for-a-repository

https://developer.github.com/v3/issues/#get-a-single-issue

Examples

if (FALSE) {

  # View open issues in a repository
  view_issues("ChadGoymer/githapi")

  # View closed issues in a repository
  view_issues("ChadGoymer/githapi", state = "closed")

  # View a single issue
  view_issue("test issue", "ChadGoymer/githapi")

  # Open a issue's page in a browser
  browse_issue("test issue", "ChadGoymer/githapi")

}