View labels within a repository

view_labels() summarises labels in a table with the properties as columns and a row for each label in the repository. It can also be used to view the labels assigned to a single issue. view_label() returns a list of all properties for a single label.

view_labels(repo, issue, n_max = 1000, ...)

view_label(label, repo, ...)

Arguments

repo
(string) The repository specified in the format: owner/repo.

issue
(string or character, optional) The number or title of the issue.

n_max
(integer, optional) Maximum number to return. Default: 1000.

...
Parameters passed to gh_page() or gh_request().

label
(string) The name of the label.

Value

view_labels() returns a tibble of label properties. view_label() returns a list of properties for a single label.

Label Properties:

name: The name of the label.

color: The color of the label in hexidecimal code.

description: The description of the label.

Details

For more details see the GitHub API documentation:

https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#list-labels-for-a-repository

https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#list-labels-for-an-issue

https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#get-a-label

Examples

if (FALSE) {

  # View all labels in a repository
  view_labels("ChadGoymer/githapi")

  # View a single label
  view_label("new-label", "ChadGoymer/githapi")

}