View milestones within a repository

view_milestones() summarises milestones in a table with the properties as columns and a row for each milestone in the repository. view_milestone() returns a list of all properties for a single milestone. browse_milestone() opens the web page for the milestone in the default browser.

view_milestones(
  repo,
  state = "open",
  sort = "due_on",
  direction = "asc",
  n_max = 1000,
  ...
)

view_milestone(milestone, repo, ...)

browse_milestone(milestone, repo, ...)

Arguments

repo
(string) The repository specified in the format: owner/repo.

state
(string, optional) The state of the milestones to filter the result. Can be either "open", "closed" or "all". Default: "open".

sort
(string, optional) The property to order the returned milestones by. Can be either "due_on" or "completeness". Default: "due_on".

direction
(string, optional) The direction of the sort. Can be either "asc" or "desc". Default: "asc".

n_max
(integer, optional) Maximum number to return. Default: 1000.

...
Parameters passed to gh_page() or gh_request().

milestone
(string or character) The number or title of the milestone.

Value

view_milestones() returns a tibble of milestone properties. view_milestone() returns a list of properties for a single milestone. browse_milestone() opens the default browser on the milestone's page and returns the URL.

Milestone Properties:

number: The number assigned to the milestone within the repository.

title: The title of the milestone.

description: The description of the milestone.

state: The state of the milestone - either "open" or "closed".

open_issues: The number of open issues within the milestone.

closed_issues: The number of closed issues within the milestone.

html_url: The URL of the milestone's web page in GitHub.

creator: The creator's login.

created_at: When the milestone was created.

updated_at: When the milestone was last updated.

due_on: When the milestone is due.

closed_at: When the milestone was closed.

Details

You can summarise all the milestones of a repository in a specified state and change the order they are returned using sort and direction.

For more details see the GitHub API documentation:

https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#list-milestones

https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#get-a-milestone

Examples

if (FALSE) {

  # View open milestones in a repository
  view_milestones("ChadGoymer/githapi")

  # View closed milestones in a repository
  view_milestones("ChadGoymer/githapi", state = "closed")

  # View a single milestone
  view_milestone(
    milestone = "test milestone",
    repo      = "ChadGoymer/githapi"
  )

  # Open a milestone's page in a browser
  browse_milestone(
    milestone = "test milestone",
    repo      = "ChadGoymer/githapi"
  )

}