Create a card in a GitHub project

This function creates a new card in the specified column in a project in GitHub. The card can either contain an existing issue or pull request or you can create a note.

create_card(
  content_id,
  content_type,
  note,
  column,
  project,
  repo,
  user,
  org,
  ...
)

Arguments

content_id
(integer, optional) The issue or pull request id associated with this card. Note: content_type is also required.

content_type
(string, optional) Whether the content_id identifies an "Issue" or a "PullRequest".

note
(string, optional) The card's note content. Not used if content_id is specified.

column
(integer or string) Either the ID or name of the column.

project
(integer or string) Either the project number or name.

repo
(string, optional) The repository specified in the format: owner/repo.

user
(string, optional) The login of the user.

org
(string, optional) The name of the organization.

...
Parameters passed to gh_request().

Value

create_card() returns a list of the card properties.

Card Properties:

id: The ID of the card.

content_id: The ID of the issue or pull request.

note: The content of a note.

archived: Whether the card has been archived.

creator: The creator of the note.

created_at: When it was created.

updated_at: When it was last updated.

Details

You can create a card in a project associated with either a repository, user or organization, by supplying them as an input, as long as you have appropriate permissions.

For more details see the GitHub API documentation:

https://developer.github.com/v3/projects/cards/#create-a-project-card

Examples

if (FALSE) {
  # Create an issue card in a repository project
  create_card(
    content_id   = 1,
    content_type = "Issue",
    column       = "Test column",
    project      = "Test project",
    repo         = "ChadGoymer/test-githapi")

  # Create a pull request card in a user's project
  create_card(
    content_id   = 2,
    content_type = "PullRequest",
    column       = "Test column",
    project      = "Test project",
    user         = "ChadGoymer")

  # Create a note card in an organization's project
  create_card(
    note    = "This is a note",
    column  = "Test column",
    project = "Test project",
    org     = "HairyCoos")
}