Update releases to new commits.

This function updates an existing releases. It can point the releases at new commits or just update the names or bodies for the releases. You can also change whether the releases are draft and whether they are pre-releases.

update_releases(repo, tags, commits = NA, names = NA, bodies = NA,
  draft = NA, prerelease = NA, token = gh_token(),
  api = getOption("github.api"), ...)

Arguments

repo
(string) The repository specified in the format: owner/repo.

tags
(character) The tag names.

commits
(character, optional) The commits to tag. Can be either a SHA or a branch name. If missing or NA then the commit is not changed. Default: NA.

names
(character, optional) The names of the releases. If missing or NA then the name is not changed. Default: NA.

bodies
(character, optional) The bodies for the releases. If missing or NA then the body is not changed. Default: NA.

draft
(logical, optional) Whether the releases are draft. Default: FALSE.

prerelease
(logical, optional) Whether the releases are a prerelease. Default: FALSE.

token
(string, optional) The personal access token for GitHub authorisation. Default: value stored in the environment variable GITHUB_TOKEN or GITHUB_PAT.

api
(string, optional) The URL of GitHub's API. Default: the value stored in the environment variable GITHUB_API or https://api.github.com.

...
Parameters passed to gh_request().

Value

A tibble describing the releases, with the following columns (see GitHub's documentation for more details):

id: The unique ID assigned by GitHub.

tag_name: The name of the tag associated with the release.

name: The name, or title, of the release.

body: The body, or description, of the release.

author_login: The author of the release.

draft: Whether the release is a draft.

prerelease: Whether the release is a pre-release.

target_commitish: The commit or branch the release was applied to.

created_at: The date and time of creation.

published_at: The date and time of publication.

assets: The names of any assets associated with the release.

zipball_url: The URL to download a zip file of the release commit.

url: The URL to view the release details on GitHub.

Details

https://developer.github.com/v3/repos/releases/#edit-a-release