Upload a directory of files and create a commit on a branch

This function uploads all the files in the specified directory path (using upload_tree()). It the creates a new commit (using create_commit()) and sets it as the head of the specified branch (using create_branches()).

upload_commit(
  branch,
  message,
  path,
  parents,
  committer,
  author,
  repo,
  replace = TRUE,
  ignore = c("\\.git", "\\.Rproj\\.user", "\\.Rhistory", "\\.RData",
    "\\.Ruserdata"),
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)

Arguments

branch
(string) The name of the branch to set the commit as head.

message
(string) The commit message

path
(string) The path to the directory, which is to be uploaded as a tree.

parents
(character, optional) The SHAs of the commits that were the parents of this commit. If not specified the commit will be written as a root commit. For a single parent, one SHA should be provided; for a merge commit, a character vector of more than one should be provided.

committer
(list, optional) The name and email address of the committer. This needs to be specified as a named list, e.g. list(name = "Bob Smith", email = "bob.smith@acme.com"). If not specified then the authenticated user is used.

author
(list, optional) The name and email address of the author. This needs to be specified as a named list, e.g. list(name = "Bob Smith", email = "bob.smith@acme.com"). If not specified then the authenticated user is used.

repo
(string) The repository specified in the format: owner/repo.

replace
(logical, optional) Whether the commit should replace the previous one or just update the files within the directory (see description). Default: TRUE.

ignore
(character) A character vector of regular expressions. If any of these are detected in a file name they are not uploaded. Default: "\\.git", "\\.Rproj\\.user", "\\.Rhistory", "\\.RData" & "\\.Ruserdata"

token
(string, optional) The personal access token for GitHub authorisation. Default: value stored in the environment variable GITHUB_TOKEN (or GITHUB_PAT) or in the R option "github.token".

api
(string, optional) The URL of GitHub's API. Default: the value stored in the environment variable GITHUB_API or in the R option "github.api".

...
Parameters passed to gh_request().

Value

A tibble describing the commit, with the following columns (see GitHub's documentation for more details):

sha: The SHA of the commit.

message: The commit message recorded in GitHub.

author_name: The author of the commit.

author_email: The author's email address.

committer_name: The committer of the commit.

committer_email: The committer's email address.

date: The date and time of the commit.

url: The URL to get the commit details from GitHub.

tree_sha: The SHA of the git tree.

tree_url: The URL to get the tree details from GitHub.

parent_sha: The SHA(s) of the parent commit(s).

parent_url: The URL(s) to get the parent's details from GitHub.

Details

The default behaviour is to treat the directory as containing the entire set of files and completely replace the previous commit with this one. This means that if a previously existing file is not in the directory it will be deleted in the commit. Alternatively, if you just want to update a few files, you can specify replace = FALSE, in which case the files in the directory will replace the ones in the previous commit, and all other files will remain unchanged. This is useful if you have a lot of files in the commit.

https://developer.github.com/v3/git/blobs/#create-a-blob

https://developer.github.com/v3/git/trees/#create-a-tree

https://developer.github.com/v3/git/commits/#create-a-commit

https://developer.github.com/v3/git/refs/#create-a-reference