View GitHub projects

view_projects() summarises projects in a table with the properties as columns and a row for each project. view_project() returns a list of all properties for a single project. browse_project() opens the web page for the project in the default browser.

view_projects(repo, user, team, org, state = "open", n_max = 1000, ...)

view_project(project, repo, user, team, org, ...)

browse_project(project, repo, user, team, org, ...)

Arguments

repo
(string, optional) The repository specified in the format: owner/repo.

user
(string, optional) The login of the user.

team
(string or integer, optional) The team ID or name.

org
(string, optional) The name of the organization.

state
(string, optional) Indicates the state of the projects to return. Can be either "open", "closed", or "all". Default: "open".

n_max
(integer, optional) Maximum number to return. Default: 1000.

...
Parameters passed to gh_page().

project
(integer or string) The number or name of the project.

Value

view_projects() returns a tibble of project properties. view_project() returns a list of properties for a single project. browse_project() opens the default browser on the project's page and returns the URL.

Project Properties:

id: The ID of the project.

number: The number of the project for the repository, user or organization.

name: The name given to the project.

body: The description given to the project.

state: Whether the project is "open" or "closed".

private: Whether the project is private (organization and team projects only).

org_permissions: The default permission for organization members (organization and team projects only).

team_permissions: The default permission for team members (team projects only).

creator: The user who created the project.

created_at: When it was created.

updated_at: When it was last updated.

html_url: The URL to view the project.

Details

You can summarise all the projects associated with either a repository, user, team or organization, by supplying them as an input.

For more details see the GitHub API documentation:

https://developer.github.com/v3/projects/#list-repository-projects

https://developer.github.com/v3/projects/#list-user-projects

https://developer.github.com/v3/projects/#list-organization-projects

https://developer.github.com/v3/teams/#review-a-team-project

https://developer.github.com/v3/projects/#get-a-project

Examples

if (FALSE) {
  # View a repository's projects
  view_projects("ChadGoymer/githapi")

  # View a user's projects
  view_projects(user = "ChadGoymer")

  # View an organization's projects
  view_projects(org = "HairyCoos")

  # View a team's projects
  view_projects(team = "HeadCoos", org = "HairyCoos")

  # View closed projects
  view_projects("ChadGoymer/githapi", state = "closed")

  # View all projects
  view_projects("ChadGoymer/githapi", state = "all")

  # View a specific repository project
  view_project("Prioritisation", repo = "ChadGoymer/githapi")

  # View a specific user project
  view_project("Test project", user = "ChadGoymer")

  # View a specific organization project
  view_project("Prioritisation", org = "HairyCoos")

  # View a specific team project
  view_project("Prioritisation", team = "HeadCoos", org = "HairyCoos")

  # Browse a specific repository project
  browse_project("Prioritisation", "ChadGoymer/githapi")

  # Browse a specific user project
  browse_project("Test project", user = "ChadGoymer")

  # Browse a specific organization project
  browse_project("Prioritisation", org = "HairyCoos")

  # Browse a specific team project
  browse_project("Prioritisation", team = "HeadCoos", org = "HairyCoos")
}